src/main/webapp/app/entities/tracked-entity-attribute/delete/tracked-entity-attribute-delete-dialog.component.ts
import { Component } from '@angular/core';
import { NgbActiveModal } from '@ng-bootstrap/ng-bootstrap';

import { ITrackedEntityAttribute } from '../tracked-entity-attribute.model';
import { TrackedEntityAttributeService } from '../service/tracked-entity-attribute.service';
import { ITEM_DELETED_EVENT } from 'app/config/navigation.constants';

@Component({
  templateUrl: './tracked-entity-attribute-delete-dialog.component.html',
})
export class TrackedEntityAttributeDeleteDialogComponent {
  trackedEntityAttribute?: ITrackedEntityAttribute;

  constructor(protected trackedEntityAttributeService: TrackedEntityAttributeService, protected activeModal: NgbActiveModal) {}

  cancel(): void {
    this.activeModal.dismiss();
  }

  confirmDelete(id: string): void {
    this.trackedEntityAttributeService.delete(id).subscribe(() => {
      this.activeModal.close(ITEM_DELETED_EVENT);
    });
  }
}
